More memory allocator things

+++ src/core/memory/vultr_memory.h
@@ -28,7 +28,7 @@ namespace Vultr
         // - Whether the block has properly initialized: 0 = Uninitialized, 1 = Initialized.
         // - Whether it is allocated or not:  0 = Freed, 1 = Allocated.
         // - The color of the block (Only valid for free blocks of memory. Used for red-black tree): 0 = Black, 1 = Red
-        u64 size = 0;
+        size_t size = 0;
 
         // Pointer to previous and next memory blocks
         MemoryBlock *prev = nullptr;
@@ -50,18 +50,18 @@ namespace Vultr
 
     /*
      * Allocate a chunk of memory from the OS and put it in a `MemoryArena`.
-     * @param u64 size: Size in bytes of how much space the memory arena should have.
+     * @param size_t size: Size in bytes of how much space the memory arena should have.
      *
      * @return MemoryArena *: The memory arena object.
      *
-     * @error The method will return nullptr if there it is unable to allocate the memory arena. This should be handled properly.
+     * @error The method will return nullptr if it is unable to allocate the memory arena. This should be handled properly.
      * */
-    MemoryArena *init_mem_arena(u64 size, u8 alignment = 16);
+    MemoryArena *init_mem_arena(size_t size, u8 alignment = 16);
 
     /*
      * Allocate a chunk of memory from a `MemoryArena`.
      * @param MemoryArena *arena: The memory arena to allocate from.
-     * @param u64 size: The size of memory to allocate.
+     * @param size_t size: The size of memory to allocate.
      *
      * @return void *: The memory that can now be used.
      *
@@ -69,7 +69,7 @@ namespace Vultr
      *
      * @no_thread_safety
      * */
-    void *mem_arena_alloc(MemoryArena *arena, u64 size);
+    void *mem_arena_alloc(MemoryArena *arena, size_t size);
 
     /*
      * Free a chunk of memory from a `MemoryArena`.
